add the RoomScene
