--- proj.win32/RoomScene.h
@@ -1,20 +1,0 @@
-#pragma once
-#ifndef __ROOMSCENE_SCENE_H__
-#define __ROOMSCENE_SCENE_H__
-
-#include"cocos2d.h"
-#include"SimpleAudioEngine.h"
-
-class RoomScene :public cocos2d::Layer
-{
-private:
-	cocos2d::Menu *createText();
-	void menuBackCallback(cocos2d::Ref* pSender);
-	void menuReadyCallback(cocos2d::Ref* pSender);
-public:
-	static cocos2d::Scene * createScene();
-	virtual bool init();
-	CREATE_FUNC(RoomScene);
-};
-#endif // !__ROOMSCENE_SCENE_H__
-
